requirements.txt
django
django-mptt
